RUE-857: split place and ownership analysis
