+++ crates/rue-air/src/sema/analysis/builtin_ops.rs
@@ -6,118 +6,6 @@
 use super::*;
 
 impl<'a> BodySema<'a> {
-    /// Give an internally synthesized borrow argument an addressable home.
-    /// Source calls normally get this from argument analysis; compiler-owned
-    /// operators also accept rvalues, so they must perform the same spill.
-    pub(crate) fn materialize_borrow_argument(
-        &mut self,
-        air: &mut Air,
-        value: AirRef,
-        ty: Type,
-        span: Span,
-        ctx: &mut AnalysisContext,
-    ) -> CompileResult<(AirRef, Vec<AirRef>)> {
-        let (value, mut prefix) = self.peel_projected_rvalue_scope(air, value);
-        if matches!(
-            air.get(value).data,
-            AirInstData::Load { .. } | AirInstData::Param { .. } | AirInstData::PlaceRead { .. }
-        ) {
-            return Ok((value, prefix));
-        }
-
-        let slot = ctx.next_slot;
-        ctx.next_slot += self.require_layout_slots(ty, span)?;
-        let live = air.add_inst(AirInst {
-            data: AirInstData::StorageLive { slot },
-            ty,
-            span,
-        });
-        let alloc = air.add_inst(AirInst {
-            data: AirInstData::Alloc { slot, init: value },
-            ty: Type::UNIT,
-            span,
-        });
-        let load = air.add_inst(AirInst {
-            data: AirInstData::Load { slot },
-            ty,
-            span,
-        });
-        prefix.extend([live, alloc]);
-        Ok((load, prefix))
-    }
-
-    pub(crate) fn wrap_value_with_temp_scope(
-        &mut self,
-        air: &mut Air,
-        value: AirRef,
-        ty: Type,
-        span: Span,
-        prefix: Vec<AirRef>,
-    ) -> CompileResult<AirRef> {
-        if prefix.is_empty() {
-            return Ok(value);
-        }
-        Ok(air.add_block(&prefix, value, ty, span)?)
-    }
-
-    /// Project the shared `{ptr, len}` text prefix from a StrBuf value.
-    ///
-    /// Runtime text consumers use this two-word ABI for every text rung. An
-    /// rvalue is spilled exactly once so both field reads share one owner.
-    pub(crate) fn project_strbuf_text_fields(
-        &mut self,
-        air: &mut Air,
-        value: AirRef,
-        ty: Type,
-        span: Span,
-        ctx: &mut AnalysisContext,
-    ) -> CompileResult<(AirRef, AirRef, Vec<AirRef>)> {
-        let struct_id = ty.as_struct().expect("StrBuf is a struct");
-        let (value, mut prefix) = self.peel_projected_rvalue_scope(air, value);
-        let (base, base_type, projections) = match air.get(value).data {
-            AirInstData::Load { slot } => (AirPlaceBase::Local(slot), ty, Vec::new()),
-            AirInstData::Param { index } => (AirPlaceBase::Param(index), ty, Vec::new()),
-            AirInstData::PlaceRead { place } => {
-                let place = air.get_place(place);
-                let projections = air.get_place_projections(place).to_vec();
-                (place.base, place.base_type, projections)
-            }
-            _ => {
-                let slot = ctx.next_slot;
-                ctx.next_slot += self.require_layout_slots(ty, span)?;
-                let live = air.add_inst(AirInst {
-                    data: AirInstData::StorageLive { slot },
-                    ty,
-                    span,
-                });
-                let alloc = air.add_inst(AirInst {
-                    data: AirInstData::Alloc { slot, init: value },
-                    ty: Type::UNIT,
-                    span,
-                });
-                prefix.extend([live, alloc]);
-                (AirPlaceBase::Local(slot), ty, Vec::new())
-            }
-        };
-
-        let fields = self.type_pool.struct_def(struct_id).fields;
-        let mut reads = Vec::with_capacity(2);
-        for field_index in 0..2u32 {
-            let mut field_projections = projections.clone();
-            field_projections.push(AirProjection::Field {
-                struct_id,
-                field_index,
-            });
-            let place = air.make_place(base, base_type, field_projections)?;
-            reads.push(air.add_inst(AirInst {
-                data: AirInstData::PlaceRead { place },
-                ty: fields[field_index as usize].ty,
-                span,
-            }));
-        }
-        Ok((reads[0], reads[1], prefix))
-    }
-
     /// Convert RIR argument mode to AIR argument mode.
     pub(super) fn convert_arg_mode(mode: RirArgMode) -> AirArgMode {
         match mode {
@@ -178,9 +66,9 @@ impl<'a> BodySema<'a> {
         // the temporary is instead dropped normally at statement end (no leak,
         // no double free). cancel_move_marker is a no-op on the Load case.
         let lhs_result = self.analyze_inst_for_projection(air, lhs, ctx)?;
-        air.cancel_move_marker(lhs_result.air_ref);
+        self.cancel_recorded_move(air, lhs_result.air_ref);
         let rhs_result = self.analyze_inst_for_projection(air, rhs, ctx)?;
-        air.cancel_move_marker(rhs_result.air_ref);
+        self.cancel_recorded_move(air, rhs_result.air_ref);
 
         // Defensive type check (HM inference already guarantees both are StrBuf
         // when we get here; this guards against error-recovery paths).
@@ -288,7 +176,7 @@ impl<'a> BodySema<'a> {
         // and cancel any move marker so a variable operand is not consumed and
         // a temporary is still dropped by its owner at statement end.
         let arg_result = self.analyze_inst_for_projection(air, arg_value, ctx)?;
-        air.cancel_move_marker(arg_result.air_ref);
+        self.cancel_recorded_move(air, arg_result.air_ref);
 
         if !self.is_strbuf(arg_result.ty)
             && !self.is_str_like(arg_result.ty)
