+++ crates/rue-air/src/sema/analysis/instructions.rs
@@ -443,361 +443,4 @@ impl<'a> BodySema<'a> {
             }
         }
     }
-
-    // ========================================================================
-    // Implementation methods for complex operations
-    // These are called by category-specific semantic modules.
-    // ========================================================================
-
-    /// Implementation for FieldSet - handles both local and parameter field assignment.
-    pub(crate) fn analyze_field_set_impl(
-        &mut self,
-        air: &mut Air,
-        base: InstRef,
-        field: Spur,
-        value: InstRef,
-        span: Span,
-        ctx: &mut AnalysisContext,
-    ) -> CompileResult<AnalysisResult> {
-        use crate::sema::analyze_ops::ProjectionInfo;
-
-        // Try to trace the base to a place
-        if let Some(mut trace) = self.try_trace_place(base, air, ctx)? {
-            // Check if the root variable was fully moved
-            if let Some(state) = ctx.moved_vars.get(&trace.root_var) {
-                if let Some(moved_span) = state.full_move {
-                    let root_name = self.interner.resolve(&trace.root_var);
-                    return Err(CompileError::new(
-                        ErrorKind::UseAfterMove(root_name.to_string()),
-                        span,
-                    )
-                    .with_label("value moved here", moved_span)
-                    .with_help(super::borrow_instead_of_move_help(root_name)));
-                }
-            }
-
-            // Writing through a field of a collection an enclosing `for` loop
-            // is iterating mutates a shared-borrowed value (spec 4.8:26,
-            // RUE-233) — E0428, like an explicit `borrow` parameter.
-            self.reject_mutate_iter_borrowed(trace.root_var, span, ctx)?;
-
-            // Check mutability
-            let root_name = self.interner.resolve(&trace.root_var).to_string();
-            if !trace.is_root_mutable {
-                // Check if this is a borrow parameter - special error message
-                if trace.is_borrow_param {
-                    return Err(CompileError::new(
-                        ErrorKind::MutateBorrowedValue {
-                            variable: root_name,
-                        },
-                        span,
-                    ));
-                }
-
-                let root_type = trace.base_type;
-                // Provide more specific error based on whether it's a param or local
-                match trace.base {
-                    AirPlaceBase::Param(_) => {
-                        return Err(self.immutable_param_assign_error(&root_name, root_type, span));
-                    }
-                    AirPlaceBase::Local(_) => {
-                        return Err(CompileError::new(
-                            ErrorKind::AssignToImmutable(root_name),
-                            span,
-                        ));
-                    }
-                }
-            }
-
-            // Add the final field projection
-            let base_type = trace.result_type();
-            let struct_id = match base_type.as_struct() {
-                Some(id) => id,
-                None => {
-                    return Err(CompileError::new(
-                        ErrorKind::FieldAccessOnNonStruct {
-                            found: base_type.safe_name_with_pool(Some(&self.type_pool)),
-                        },
-                        span,
-                    ));
-                }
-            };
-
-            let struct_def = self.type_pool.struct_def(struct_id);
-            let field_name_str = self.interner.resolve(&field).to_string();
-
-            let (field_index, struct_field) =
-                struct_def.find_field(&field_name_str).ok_or_compile_error(
-                    ErrorKind::UnknownField {
-                        struct_name: struct_def.name.clone(),
-                        field_name: field_name_str.clone(),
-                    },
-                    span,
-                )?;
-
-            let field_type = struct_field.ty;
-
-            // Add the field projection to the trace
-            trace.projections.push(ProjectionInfo {
-                proj: AirProjection::Field {
-                    struct_id,
-                    field_index: field_index as u32,
-                },
-                result_type: field_type,
-                field_name: Some(field),
-                const_index: None,
-                index_segment: None,
-            });
-
-            // A write through an element of a partially moved array
-            // (`xs[0].f = ...` after an element of `xs` moved out) is
-            // rejected (RUE-186, E0480), like a direct element write.
-            self.reject_write_into_partially_moved_array(&trace, ctx, span)?;
-
-            // Analyze the value
-            let value_result = self.analyze_inst(air, value, ctx)?;
-
-            // RUE-387: writing a live linear value's field would silently drop
-            // the old field value. Legal only when that exact field path was
-            // proven moved-out on every path (`consume(o.f); o.f = ...`); a
-            // dynamic index anywhere in the chain can never prove it.
-            let discharged = !trace.has_untrackable_index()
-                && self.place_linear_discharged(
-                    field_type,
-                    trace.root_var,
-                    &trace.field_path(),
-                    span,
-                    ctx,
-                );
-            self.check_linear_overwrite(field_type, discharged, false, span)?;
-
-            // The write reinitializes its destination: the assigned path
-            // (and any moved sub-paths under it) is no longer moved, so
-            // `o.f = ...` after `consume(o.f)` makes `o.f` usable again.
-            // Writes through an index projection are skipped: a partially
-            // moved array is already rejected outright above (E0480), and
-            // per-element write re-arm is unsupported (the sema path and the
-            // runtime drop flags would disagree), so the whole array must be
-            // reinitialized instead — never a single `arr[0].f` write.
-            if !trace
-                .projections
-                .iter()
-                .any(|p| matches!(p.proj, AirProjection::Index { .. }))
-            {
-                let assigned_path = trace.field_path();
-                if let Some(state) = ctx.moved_vars.get_mut(&trace.root_var) {
-                    state.mark_path_reinitialized(&assigned_path);
-                    if state.is_empty() {
-                        ctx.moved_vars.remove(&trace.root_var);
-                    }
-                }
-            }
-
-            // Emit PlaceWrite instruction
-            let place_ref = Self::build_place_ref(air, &trace)?;
-            let air_ref = air.add_inst(AirInst {
-                data: AirInstData::PlaceWrite {
-                    place: place_ref,
-                    value: value_result.air_ref,
-                },
-                ty: Type::UNIT,
-                span,
-            });
-            return Ok(AnalysisResult::new(air_ref, Type::UNIT));
-        }
-
-        // Fallback: base is not a place (e.g., function call result)
-        // This shouldn't normally happen for valid assignment targets
-        Err(CompileError::new(ErrorKind::InvalidAssignmentTarget, span))
-    }
-
-    /// Implementation for IndexSet - handles both local and parameter array index assignment.
-    pub(crate) fn analyze_index_set_impl(
-        &mut self,
-        air: &mut Air,
-        base: InstRef,
-        index: InstRef,
-        value: InstRef,
-        span: Span,
-        ctx: &mut AnalysisContext,
-    ) -> CompileResult<AnalysisResult> {
-        use crate::sema::analyze_ops::ProjectionInfo;
-
-        // Try to trace the base to a place
-        if let Some(mut trace) = self.try_trace_place(base, air, ctx)? {
-            // Check if the root variable was fully moved
-            if let Some(state) = ctx.moved_vars.get(&trace.root_var) {
-                if let Some(moved_span) = state.full_move {
-                    let root_name = self.interner.resolve(&trace.root_var);
-                    return Err(CompileError::new(
-                        ErrorKind::UseAfterMove(root_name.to_string()),
-                        span,
-                    )
-                    .with_label("value moved here", moved_span)
-                    .with_help(super::borrow_instead_of_move_help(root_name)));
-                }
-            }
-
-            // Writing an element of a collection an enclosing `for` loop is
-            // iterating mutates a shared-borrowed value (spec 4.8:26,
-            // RUE-233) — E0428, like an explicit `borrow` parameter.
-            self.reject_mutate_iter_borrowed(trace.root_var, span, ctx)?;
-
-            // Check mutability
-            let root_name = self.interner.resolve(&trace.root_var).to_string();
-            if !trace.is_root_mutable {
-                // Check if this is a borrow parameter - special error message
-                if trace.is_borrow_param {
-                    return Err(CompileError::new(
-                        ErrorKind::MutateBorrowedValue {
-                            variable: root_name,
-                        },
-                        span,
-                    ));
-                }
-
-                let root_type = trace.base_type;
-                match trace.base {
-                    AirPlaceBase::Param(_) => {
-                        return Err(self.immutable_param_assign_error(&root_name, root_type, span));
-                    }
-                    AirPlaceBase::Local(_) => {
-                        return Err(CompileError::new(
-                            ErrorKind::AssignToImmutable(root_name),
-                            span,
-                        ));
-                    }
-                }
-            }
-
-            // Get array type info from the trace
-            let base_type = trace.result_type();
-            let (_array_type_id, elem_type, array_len) = match base_type.as_array() {
-                Some(id) => {
-                    let (elem, len) = self.type_pool.array_def(id);
-                    (id, elem, len)
-                }
-                None => {
-                    return Err(CompileError::new(
-                        ErrorKind::IndexOnNonArray {
-                            found: base_type.safe_name_with_pool(Some(&self.type_pool)),
-                        },
-                        span,
-                    ));
-                }
-            };
-
-            // Analyze index. Index must be an integer type (signed or
-            // unsigned) per spec 7.1:7; negative/out-of-range runtime indices
-            // trap at runtime via the bounds check (RUE-81).
-            let index_result = self.analyze_inst(air, index, ctx)?;
-            if !index_result.ty.is_integer() && !index_result.ty.is_error() {
-                return Err(CompileError::new(
-                    ErrorKind::TypeMismatch {
-                        expected: "integer type".to_string(),
-                        found: index_result.ty.safe_name_with_pool(Some(&self.type_pool)),
-                    },
-                    self.rir.get(index).span,
-                ));
-            }
-
-            // Compile-time bounds check for constant indices, evaluated at the
-            // index's resolved operand types so an overflowing index expression
-            // is a compile-time error, not a folded runtime panic (RUE-234).
-            if let Some(const_index) = self.try_get_const_index_checked(index, ctx)? {
-                if const_index < 0 || const_index >= array_len as i128 {
-                    return Err(CompileError::new(
-                        ErrorKind::IndexOutOfBounds {
-                            index: const_index,
-                            length: array_len,
-                        },
-                        self.rir.get(index).span,
-                    ));
-                }
-            }
-
-            // Add the index projection. A non-negative constant index carries
-            // its element path segment so field_path nests through it (RUE-279).
-            let const_index = self.try_get_const_index(index);
-            let index_segment = match const_index {
-                Some(k) if k >= 0 => Some(index_path_segment(self.interner, k as u64)),
-                _ => None,
-            };
-            trace.projections.push(ProjectionInfo {
-                proj: AirProjection::Index {
-                    array_type: base_type,
-                    index: index_result.air_ref,
-                },
-                result_type: elem_type,
-                field_name: None,
-                const_index,
-                index_segment,
-            });
-
-            // Writing into an array with moved-out elements is rejected
-            // (RUE-186, E0480): the write can't re-arm per-element ownership.
-            self.reject_write_into_partially_moved_array(&trace, ctx, span)?;
-
-            // Analyze the value
-            let value_result = self.analyze_inst(air, value, ctx)?;
-
-            // RUE-387: writing a live linear value into an array element would
-            // silently drop the old element. Legal only when that exact
-            // constant-index element was proven moved-out on every path
-            // (`arr[0] = f(arr[0])`); a runtime index can never prove which
-            // element is live, so it is always rejected (repro 3).
-            let discharged = !trace.has_untrackable_index()
-                && self.place_linear_discharged(
-                    elem_type,
-                    trace.root_var,
-                    &trace.field_path(),
-                    span,
-                    ctx,
-                );
-            self.check_linear_overwrite(elem_type, discharged, false, span)?;
-
-            // The write reinitializes its destination element: the assigned
-            // path is no longer moved, so `arr[0] = arr[0]` un-marks the
-            // move-out the RHS `arr[0]` just recorded and leaves the element
-            // usable again (spec 3.8:55). This mirrors `analyze_field_set_impl`,
-            // which reinitializes an assigned field path. Only a direct,
-            // constant-index element of a root array is reinitialized: that is
-            // exactly the shape `record_element_move_out` tracks per element
-            // (`projections == [Index]` with a known index), so a nested or
-            // dynamic index — which was never recorded as a per-element move —
-            // is conservatively left alone (RUE-228).
-            if let [
-                ProjectionInfo {
-                    const_index: Some(k),
-                    ..
-                },
-            ] = trace.projections.as_slice()
-            {
-                if *k >= 0 {
-                    let elem_path = vec![index_path_segment(self.interner, *k as u64)];
-                    if let Some(state) = ctx.moved_vars.get_mut(&trace.root_var) {
-                        state.mark_path_reinitialized(&elem_path);
-                        if state.is_empty() {
-                            ctx.moved_vars.remove(&trace.root_var);
-                        }
-                    }
-                }
-            }
-
-            // Emit PlaceWrite instruction
-            let place_ref = Self::build_place_ref(air, &trace)?;
-            let air_ref = air.add_inst(AirInst {
-                data: AirInstData::PlaceWrite {
-                    place: place_ref,
-                    value: value_result.air_ref,
-                },
-                ty: Type::UNIT,
-                span,
-            });
-            return Ok(AnalysisResult::new(air_ref, Type::UNIT));
-        }
-
-        // Fallback: base is not a place
-        Err(CompileError::new(ErrorKind::InvalidAssignmentTarget, span))
-    }
 }
